Numerous updates related to the Fst/dDAF cluster analysis, v12, and stability

+++ lib/create_schema_Alignment.sql
@@ -5,10 +5,10 @@ chain VARCHAR(10),
 chain_seqid INT,
 transcript VARCHAR(50),
 trans_seqid INT,
-al_id MEDIUMINT NOT NULL AUTO_INCREMENT, # Unique, direct-reference key
-as_id MEDIUMINT, # AlignmentScore direct-reference key
-res_id MEDIUMINT, # Residue direct-reference key
-tr_id MEDIUMINT, # Transcript direct-reference key
+al_id BIGINT NOT NULL AUTO_INCREMENT, # Unique, direct-reference key
+as_id BIGINT, # AlignmentScore direct-reference key
+res_id BIGINT, # Residue direct-reference key
+tr_id BIGINT, # Transcript direct-reference key
 PRIMARY KEY(label,structid,chain,chain_seqid,transcript,trans_seqid),
 KEY(al_id),
 KEY(as_id),
