+++ lib/create_schema_Chain.sql
@@ -8,8 +8,8 @@ unp VARCHAR(50),
 offset INT,
 hybrid TINYINT,
 sequence TEXT,
-ch_id MEDIUMINT NOT NULL AUTO_INCREMENT, # Unique, direct-reference key
-str_id MEDIUMINT, # Structure/Model direct-reference key
+ch_id BIGINT NOT NULL AUTO_INCREMENT, # Unique, direct-reference key
+str_id BIGINT, # Structure/Model direct-reference key
 PRIMARY KEY(label,structid,biounit,model,chain),
 KEY(ch_id),
 KEY(str_id),
