+++ lib/create_schema_Model.sql
@@ -12,7 +12,7 @@ zdope DOUBLE,
 pdbid VARCHAR(50),
 chain VARCHAR(10),
 identity DOUBLE,
-str_id MEDIUMINT NOT NULL AUTO_INCREMENT, # Unique, direct-reference key
+str_id BIGINT NOT NULL AUTO_INCREMENT, # Unique, direct-reference key
 PRIMARY KEY(label,modelid),
 KEY(str_id),
 KEY(label,pdbid),
